// src/enums/EnumMQTTCmd.ts
export enum EnumMQTTCmd {
    GET_DEFAULT_ACL_ACCESS = "getDefaultACLAccess",
    SET_DEFAULT_ACL_ACCESS = "setDefaultACLAccess",
    GET_ANONYMOUS_GROUP = "getAnonymousGroup",
    SET_ANONYMOUS_GROUP = "setAnonymousGroup",
    CREATE_CLIENT = "createClient",
    DELETE_CLIENT = "deleteClient",
    SET_CLIENT_PASSWORD = "setClientPassword",
    SET_CLIENT_ID = "setClientId",
    ADD_CLIENT_ROLE = "addClientRole",
    REMOVE_CLIENT_ROLE = "removeClientRole",
    GET_CLIENT = "getClient",
    LIST_CLIENTS = "listClients",
    ENABLE_CLIENT = "enableClient",
    DISABLE_CLIENT = "disableClient",
    CREATE_GROUP = "createGroup",
    DELETE_GROUP = "deleteGroup",
    ADD_GROUP_ROLE = "addGroupRole",
    REMOVE_GROUP_ROLE = "removeGroupRole",
    ADD_GROUP_CLIENT = "addGroupClient",
    REMOVE_GROUP_CLIENT = "removeGroupClient",
    GET_GROUP = "getGroup",
    LIST_GROUPS = "listGroups",
    CREATE_ROLE = "createRole",
    DELETE_ROLE = "deleteRole",
    ADD_ROLE_ACL = "addRoleACL",
    REMOVE_ROLE_ACL = "removeRoleACL",
    GET_ROLE = "getRole",
    LIST_ROLES = "listRoles"
}